Added check for pool (#1463) (minor)
### Added
- New function for findbusinessdata `get_case_pool` which returns the pool connected to a case if any.

### Changed

- The command `cg workflow fastq store-available` now has an extra if clause where it only stores cases that are pools or non-rml cases where all samples have been sequenced and pass QC.
