+++ cg/cli/workflow/fastq/base.py
@@ -48,4 +48,5 @@ def store_available_fastq_analysis(context: click.Context, dry_run: bool = False
     """Creates an analysis object in status-db for all fastq cases to be delivered"""
     status_db: Store = context.obj.status_db
     for case in status_db.cases_to_analyze(pipeline=Pipeline.FASTQ, threshold=False):
-        context.invoke(store_fastq_analysis, case_id=case.internal_id, dry_run=dry_run)
+        if case.all_samples_pass_qc or status_db.is_pool(case_id=case.internal_id):
+            context.invoke(store_fastq_analysis, case_id=case.internal_id, dry_run=dry_run)
